+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [v0.16.0] - ??
+
+### Added
+
+- New endpoints for temporary participants.
+- New endpoint to be able to download response export in a JSON format.
+
+### Changed
+
+- To allow workflows where registration is triggered from a survey, study flow methods don't require the account to be confirmed yet.
+
 ## [v0.15.1] - 2021-07-28
 
 - Update dependency for messaging service, to use recent API.